+++ packages/brick_offline_first/example/lib/app/adapters/kitchen_sink_adapter.g.dart
@@ -9,36 +9,40 @@ Future<KitchenSink> _$KitchenSinkFromRest(Map<String, dynamic> data,
       anyInt: data['any_int'] as int,
       anyDouble: data['any_double'] as double,
       anyNum: data['any_num'] as num,
-      anyDateTime: data['any_date_time'] == null
-          ? null
-          : DateTime.tryParse(data['any_date_time'] as String),
+      anyDateTime:
+          data['any_date_time'] == null ? null : DateTime.tryParse(data['any_date_time'] as String),
       anyBool: data['any_bool'] as bool,
       anyMap: data['any_map'],
-      enumFromIndex: data['enum_from_index'] is int
-          ? AnyEnum.values[data['enum_from_index'] as int]
-          : null,
+      enumFromIndex:
+          data['enum_from_index'] is int ? AnyEnum.values[data['enum_from_index'] as int] : null,
       anyList: data['any_list']?.toList()?.cast<int>() ?? <int>[],
       anySet: data['any_set']?.toSet()?.cast<int>() ?? <int>{},
-      offlineFirstModel: await MountyAdapter().fromRest(data['offline_first_model'],
-          provider: provider, repository: repository),
-      listOfflineFirstModel: await Future.wait<Mounty>(
-          data['list_offline_first_model']?.map((d) => MountyAdapter().fromRest(d, provider: provider, repository: repository))?.toList()?.cast<Future<Mounty>>() ??
-              []),
+      offlineFirstModel: await MountyAdapter()
+          .fromRest(data['offline_first_model'], provider: provider, repository: repository),
+      listOfflineFirstModel: await Future.wait<Mounty>(data['list_offline_first_model']
+              ?.map((d) => MountyAdapter().fromRest(d, provider: provider, repository: repository))
+              ?.toList()
+              ?.cast<Future<Mounty>>() ??
+          []),
       setOfflineFirstModel:
           (await Future.wait<Mounty>(data['set_offline_first_model']?.map((d) => MountyAdapter().fromRest(d, provider: provider, repository: repository))?.toSet()?.cast<Future<Mounty>>() ?? []))
               ?.toSet(),
-      futureOfflineFirstModel: MountyAdapter().fromRest(
-          data['future_offline_first_model'],
-          provider: provider,
-          repository: repository),
+      futureOfflineFirstModel: MountyAdapter()
+          .fromRest(data['future_offline_first_model'], provider: provider, repository: repository),
       futureListOfflineFirstModel: data['future_list_offline_first_model']
-          ?.map((d) =>
-              MountyAdapter().fromRest(d, provider: provider, repository: repository))
+          ?.map((d) => MountyAdapter().fromRest(d, provider: provider, repository: repository))
           ?.toList()
           ?.cast<Future<Mounty>>(),
-      futureSetOfflineFirstModel: (data['future_set_offline_first_model']?.map((d) => MountyAdapter().fromRest(d, provider: provider, repository: repository))?.toSet()?.cast<Future<Mounty>>())?.toSet(),
+      futureSetOfflineFirstModel: (data['future_set_offline_first_model']
+              ?.map((d) => MountyAdapter().fromRest(d, provider: provider, repository: repository))
+              ?.toSet()
+              ?.cast<Future<Mounty>>())
+          ?.toSet(),
       offlineFirstSerdes: Hat.fromRest(data['offline_first_serdes']),
-      listOfflineFirstSerdes: data['list_offline_first_serdes'].map((c) => Hat.fromRest(c as Map<String, dynamic>))?.toList()?.cast<Hat>(),
+      listOfflineFirstSerdes: data['list_offline_first_serdes']
+          .map((c) => Hat.fromRest(c as Map<String, dynamic>))
+          ?.toList()
+          ?.cast<Hat>(),
       setOfflineFirstSerdes: data['set_offline_first_serdes'].map((c) => Hat.fromRest(c as Map<String, dynamic>))?.toSet()?.cast<Hat>(),
       restAnnotationName: data['restAnnotationOtherName'] as String,
       restAnnotationDefaultValue: data['rest_annotation_default_value'] as String ?? "a default value",
@@ -67,35 +71,27 @@ Future<Map<String, dynamic>> _$KitchenSinkToRest(KitchenSink instance,
     'any_date_time': instance.anyDateTime?.toIso8601String(),
     'any_bool': instance.anyBool,
     'any_map': instance.anyMap,
-    'enum_from_index': instance.enumFromIndex != null
-        ? AnyEnum.values.indexOf(instance.enumFromIndex)
-        : null,
+    'enum_from_index':
+        instance.enumFromIndex != null ? AnyEnum.values.indexOf(instance.enumFromIndex) : null,
     'any_list': instance.anyList,
     'any_set': instance.anySet,
-    'offline_first_model':
-        await MountyAdapter().toRest(instance.offlineFirstModel),
-    'list_offline_first_model': await Future.wait<Map<String, dynamic>>(instance
-            .listOfflineFirstModel
-            ?.map((s) => MountyAdapter().toRest(s))
-            ?.toList() ??
-        []),
-    'set_offline_first_model': await Future.wait<Map<String, dynamic>>(instance
-            .setOfflineFirstModel
-            ?.map((s) => MountyAdapter().toRest(s))
-            ?.toList() ??
-        []),
+    'offline_first_model': await MountyAdapter().toRest(instance.offlineFirstModel),
+    'list_offline_first_model': await Future.wait<Map<String, dynamic>>(
+        instance.listOfflineFirstModel?.map((s) => MountyAdapter().toRest(s))?.toList() ?? []),
+    'set_offline_first_model': await Future.wait<Map<String, dynamic>>(
+        instance.setOfflineFirstModel?.map((s) => MountyAdapter().toRest(s))?.toList() ?? []),
     'future_offline_first_model':
         await MountyAdapter().toRest((await instance.futureOfflineFirstModel)),
-    'future_list_offline_first_model': await Future.wait<Map<String, dynamic>>(
-        instance.futureListOfflineFirstModel
-                ?.map((s) async => MountyAdapter().toRest((await s)))
-                ?.toList() ??
-            []),
-    'future_set_offline_first_model': await Future.wait<Map<String, dynamic>>(
-        instance.futureSetOfflineFirstModel
-                ?.map((s) async => MountyAdapter().toRest((await s)))
-                ?.toList() ??
-            []),
+    'future_list_offline_first_model': await Future.wait<Map<String, dynamic>>(instance
+            .futureListOfflineFirstModel
+            ?.map((s) async => MountyAdapter().toRest((await s)))
+            ?.toList() ??
+        []),
+    'future_set_offline_first_model': await Future.wait<Map<String, dynamic>>(instance
+            .futureSetOfflineFirstModel
+            ?.map((s) async => MountyAdapter().toRest((await s)))
+            ?.toList() ??
+        []),
     'offline_first_serdes': instance.offlineFirstSerdes?.toRest(),
     'list_offline_first_serdes':
         instance.listOfflineFirstSerdes?.map((Hat c) => c?.toRest())?.toList(),
@@ -106,8 +102,7 @@ Future<Map<String, dynamic>> _$KitchenSinkToRest(KitchenSink instance,
     'rest_annotation_nullable': instance.restAnnotationNullable,
     'rest_annotation_ignore_from': instance.restAnnotationIgnoreFrom,
     'rest_annotation_from_generator': instance.restAnnotationFromGenerator,
-    'rest_annotation_to_generator':
-        instance.restAnnotationToGenerator.toString(),
+    'rest_annotation_to_generator': instance.restAnnotationToGenerator.toString(),
     'enum_from_string': instance.enumFromString?.toString()?.split('.')?.last,
     'sqlite_annotation_nullable': instance.sqliteAnnotationNullable,
     'sqlite_annotation_default_value': instance.sqliteAnnotationDefaultValue,
@@ -121,14 +116,11 @@ Future<Map<String, dynamic>> _$KitchenSinkToRest(KitchenSink instance,
 }
 
 Future<KitchenSink> _$KitchenSinkFromSqlite(Map<String, dynamic> data,
-    {SqliteProvider provider,
-    OfflineFirstWithRestRepository repository}) async {
+    {SqliteProvider provider, OfflineFirstWithRestRepository repository}) async {
   return KitchenSink(
-      anyString:
-          data['any_string'] == null ? null : data['any_string'] as String,
+      anyString: data['any_string'] == null ? null : data['any_string'] as String,
       anyInt: data['any_int'] == null ? null : data['any_int'] as int,
-      anyDouble:
-          data['any_double'] == null ? null : data['any_double'] as double,
+      anyDouble: data['any_double'] == null ? null : data['any_double'] as double,
       anyNum: data['any_num'] == null ? null : data['any_num'] as num,
       anyDateTime: data['any_date_time'] == null
           ? null
@@ -139,28 +131,22 @@ Future<KitchenSink> _$KitchenSinkFromSqlite(Map<String, dynamic> data,
       anyMap: data['any_map'] == null ? null : jsonDecode(data['any_map']),
       enumFromIndex: data['enum_from_index'] == null
           ? null
-          : (data['enum_from_index'] > -1
-              ? AnyEnum.values[data['enum_from_index'] as int]
-              : null),
-      anyList: data['any_list'] == null
-          ? null
-          : jsonDecode(data['any_list'])?.toList()?.cast<int>(),
-      anySet: data['any_set'] == null
-          ? null
-          : jsonDecode(data['any_set'])?.toSet()?.cast<int>(),
+          : (data['enum_from_index'] > -1 ? AnyEnum.values[data['enum_from_index'] as int] : null),
+      anyList:
+          data['any_list'] == null ? null : jsonDecode(data['any_list'])?.toList()?.cast<int>(),
+      anySet: data['any_set'] == null ? null : jsonDecode(data['any_set'])?.toSet()?.cast<int>(),
       offlineFirstModel: data['offline_first_model_Mounty_brick_id'] == null
           ? null
           : (data['offline_first_model_Mounty_brick_id'] > -1
               ? (await repository?.getAssociation<Mounty>(
-                  Query.where('primaryKey',
-                      data['offline_first_model_Mounty_brick_id'] as int,
+                  Query.where('primaryKey', data['offline_first_model_Mounty_brick_id'] as int,
                       limit1: true),
                 ))
                   ?.first
               : null),
       listOfflineFirstModel: (await provider
-              ?.rawQuery('SELECT `Mounty_brick_id` FROM `_brick_KitchenSink_list_offline_first_model` WHERE KitchenSink_brick_id = ?',
-                  [data['_brick_id'] as int])?.then((results) {
+              ?.rawQuery('SELECT `Mounty_brick_id` FROM `_brick_KitchenSink_list_offline_first_model` WHERE KitchenSink_brick_id = ?', [data['_brick_id'] as int])?.then(
+                  (results) {
         final ids = results.map((r) => (r ?? {})['Mounty_brick_id']);
         return Future.wait<Mounty>(ids.map((primaryKey) => repository
             ?.getAssociation<Mounty>(
@@ -182,36 +168,40 @@ Future<KitchenSink> _$KitchenSinkFromSqlite(Map<String, dynamic> data,
       }))
           .toSet()
           .cast<Mounty>(),
-      futureOfflineFirstModel:
-          data['future_offline_first_model_Mounty_brick_id'] == null
-              ? null
-              : (data['future_offline_first_model_Mounty_brick_id'] > -1
-                  ? repository
-                      ?.getAssociation<Mounty>(
-                        Query.where(
-                            'primaryKey',
-                            data['future_offline_first_model_Mounty_brick_id']
-                                as int,
-                            limit1: true),
-                      )
-                      ?.then((r) => (r?.isEmpty ?? true) ? null : r.first)
-                  : null),
-      futureListOfflineFirstModel: await provider
-          ?.rawQuery('SELECT `Mounty_brick_id` FROM `_brick_KitchenSink_future_list_offline_first_model` WHERE KitchenSink_brick_id = ?', [data['_brick_id'] as int])?.then((results) {
+      futureOfflineFirstModel: data['future_offline_first_model_Mounty_brick_id'] == null
+          ? null
+          : (data['future_offline_first_model_Mounty_brick_id'] > -1
+              ? repository
+                  ?.getAssociation<Mounty>(
+                    Query.where(
+                        'primaryKey', data['future_offline_first_model_Mounty_brick_id'] as int,
+                        limit1: true),
+                  )
+                  ?.then((r) => (r?.isEmpty ?? true) ? null : r.first)
+              : null),
+      futureListOfflineFirstModel: await provider?.rawQuery(
+          'SELECT `Mounty_brick_id` FROM `_brick_KitchenSink_future_list_offline_first_model` WHERE KitchenSink_brick_id = ?',
+          [data['_brick_id'] as int])?.then((results) {
         final ids = results.map((r) => (r ?? {})['Mounty_brick_id']);
-        return ids.map((primaryKey) => repository
-            ?.getAssociation<Mounty>(
-              Query.where('primaryKey', primaryKey, limit1: true),
-            )
-            ?.then((r) => (r?.isEmpty ?? true) ? null : r.first)).toList().cast<Future<Mounty>>();
+        return ids
+            .map((primaryKey) => repository
+                ?.getAssociation<Mounty>(
+                  Query.where('primaryKey', primaryKey, limit1: true),
+                )
+                ?.then((r) => (r?.isEmpty ?? true) ? null : r.first))
+            .toList()
+            .cast<Future<Mounty>>();
       }),
       futureSetOfflineFirstModel: await provider?.rawQuery('SELECT `Mounty_brick_id` FROM `_brick_KitchenSink_future_set_offline_first_model` WHERE KitchenSink_brick_id = ?', [data['_brick_id'] as int])?.then((results) {
         final ids = results.map((r) => (r ?? {})['Mounty_brick_id']);
-        return ids.map((primaryKey) => repository
-            ?.getAssociation<Mounty>(
-              Query.where('primaryKey', primaryKey, limit1: true),
-            )
-            ?.then((r) => (r?.isEmpty ?? true) ? null : r.first)).toSet().cast<Future<Mounty>>();
+        return ids
+            .map((primaryKey) => repository
+                ?.getAssociation<Mounty>(
+                  Query.where('primaryKey', primaryKey, limit1: true),
+                )
+                ?.then((r) => (r?.isEmpty ?? true) ? null : r.first))
+            .toSet()
+            .cast<Future<Mounty>>();
       }),
       offlineFirstSerdes: data['offline_first_serdes'] == null ? null : Hat.fromSqlite(data['offline_first_serdes'] as String),
       listOfflineFirstSerdes: data['list_offline_first_serdes'] == null ? null : jsonDecode(data['list_offline_first_serdes']).map((c) => Hat.fromSqlite(c as String))?.toList()?.cast<Hat>(),
@@ -235,8 +225,7 @@ Future<KitchenSink> _$KitchenSinkFromSqlite(Map<String, dynamic> data,
           ? null
           : (data['offline_first_where_Mounty_brick_id'] > -1
               ? (await repository?.getAssociation<Mounty>(
-                  Query.where('primaryKey',
-                      data['offline_first_where_Mounty_brick_id'] as int,
+                  Query.where('primaryKey', data['offline_first_where_Mounty_brick_id'] as int,
                       limit1: true),
                 ))
                   ?.first
@@ -245,8 +234,7 @@ Future<KitchenSink> _$KitchenSinkFromSqlite(Map<String, dynamic> data,
 }
 
 Future<Map<String, dynamic>> _$KitchenSinkToSqlite(KitchenSink instance,
-    {SqliteProvider provider,
-    OfflineFirstWithRestRepository repository}) async {
+    {SqliteProvider provider, OfflineFirstWithRestRepository repository}) async {
   return {
     'any_string': instance.anyString,
     'any_int': instance.anyInt,
@@ -258,30 +246,22 @@ Future<Map<String, dynamic>> _$KitchenSinkToSqlite(KitchenSink instance,
     'enum_from_index': AnyEnum.values.indexOf(instance.enumFromIndex),
     'any_list': jsonEncode(instance.anyList ?? []),
     'any_set': jsonEncode(instance.anySet?.toList() ?? []),
-    'offline_first_model_Mounty_brick_id':
-        instance.offlineFirstModel?.primaryKey ??
-            await provider?.upsert<Mounty>(instance.offlineFirstModel,
-                repository: repository),
-    'set_offline_first_model':
-        jsonEncode(instance.setOfflineFirstModel?.toList() ?? []),
+    'offline_first_model_Mounty_brick_id': instance.offlineFirstModel?.primaryKey ??
+        await provider?.upsert<Mounty>(instance.offlineFirstModel, repository: repository),
+    'set_offline_first_model': jsonEncode(instance.setOfflineFirstModel?.toList() ?? []),
     'future_offline_first_model_Mounty_brick_id':
         (await instance.futureOfflineFirstModel)?.primaryKey ??
-            await provider?.upsert<Mounty>(
-                (await instance.futureOfflineFirstModel),
+            await provider?.upsert<Mounty>((await instance.futureOfflineFirstModel),
                 repository: repository),
     'future_set_offline_first_model':
         jsonEncode(instance.futureSetOfflineFirstModel?.toList() ?? []),
     'offline_first_serdes': instance.offlineFirstSerdes?.toSqlite(),
-    'list_offline_first_serdes': jsonEncode(instance.listOfflineFirstSerdes
-            ?.map((Hat c) => c?.toSqlite())
-            ?.toList()
-            ?.cast<String>() ??
-        []),
-    'set_offline_first_serdes': jsonEncode(instance.setOfflineFirstSerdes
-            ?.map((Hat c) => c?.toSqlite())
-            ?.toList()
-            ?.cast<String>() ??
-        []),
+    'list_offline_first_serdes': jsonEncode(
+        instance.listOfflineFirstSerdes?.map((Hat c) => c?.toSqlite())?.toList()?.cast<String>() ??
+            []),
+    'set_offline_first_serdes': jsonEncode(
+        instance.setOfflineFirstSerdes?.map((Hat c) => c?.toSqlite())?.toList()?.cast<String>() ??
+            []),
     'rest_annotation_name': instance.restAnnotationName,
     'rest_annotation_default_value': instance.restAnnotationDefaultValue,
     'rest_annotation_nullable': instance.restAnnotationNullable,
@@ -294,14 +274,11 @@ Future<Map<String, dynamic>> _$KitchenSinkToSqlite(KitchenSink instance,
     'sqlite_annotation_nullable': instance.sqliteAnnotationNullable,
     'sqlite_annotation_default_value': instance.sqliteAnnotationDefaultValue,
     'sqlite_annotation_from_generator': instance.sqliteAnnotationFromGenerator,
-    'sqlite_annotation_to_generator':
-        instance.sqliteAnnotationToGenerator.toString(),
+    'sqlite_annotation_to_generator': instance.sqliteAnnotationToGenerator.toString(),
     'sqlite_annotation_unique': instance.sqliteAnnotationUnique,
     'custom column name': instance.sqliteAnnotationName,
-    'offline_first_where_Mounty_brick_id':
-        instance.offlineFirstWhere?.primaryKey ??
-            await provider?.upsert<Mounty>(instance.offlineFirstWhere,
-                repository: repository)
+    'offline_first_where_Mounty_brick_id': instance.offlineFirstWhere?.primaryKey ??
+        await provider?.upsert<Mounty>(instance.offlineFirstWhere, repository: repository)
   };
 }
 
@@ -530,15 +507,14 @@ class KitchenSinkAdapter extends OfflineFirstWithRestAdapter<KitchenSink> {
       'association': true,
     }
   };
-  Future<int> primaryKeyByUniqueColumns(
-      KitchenSink instance, DatabaseExecutor executor) async {
+  Future<int> primaryKeyByUniqueColumns(KitchenSink instance, DatabaseExecutor executor) async {
     final results = await executor.rawQuery('''
         SELECT * FROM `KitchenSink` WHERE sqlite_annotation_unique = ? LIMIT 1''',
         [instance.sqliteAnnotationUnique]);
 
     // SQFlite returns [{}] when no results are found
-    if (results?.isEmpty == true ||
-        (results?.length == 1 && results?.first?.isEmpty == true)) return null;
+    if (results?.isEmpty == true || (results?.length == 1 && results?.first?.isEmpty == true))
+      return null;
 
     return results.first['_brick_id'];
   }
@@ -547,61 +523,49 @@ class KitchenSinkAdapter extends OfflineFirstWithRestAdapter<KitchenSink> {
   Future<void> afterSave(instance, {provider, repository}) async {
     if (instance.primaryKey != null) {
       await Future.wait<int>(instance.listOfflineFirstModel?.map((s) async {
-        final id = s?.primaryKey ??
-            await provider?.upsert<Mounty>(s, repository: repository);
+        final id = s?.primaryKey ?? await provider?.upsert<Mounty>(s, repository: repository);
         return await provider?.rawInsert(
-            'INSERT OR REPLACE INTO `_brick_KitchenSink_list_offline_first_model` (`KitchenSink_brick_id`, `Mounty_brick_id`) VALUES (?, ?)',
+            'INSERT OR IGNORE INTO `_brick_KitchenSink_list_offline_first_model` (`KitchenSink_brick_id`, `Mounty_brick_id`) VALUES (?, ?)',
             [instance.primaryKey, id]);
       }));
     }
 
     if (instance.primaryKey != null) {
       await Future.wait<int>(instance.setOfflineFirstModel?.map((s) async {
-        final id = s?.primaryKey ??
-            await provider?.upsert<Mounty>(s, repository: repository);
+        final id = s?.primaryKey ?? await provider?.upsert<Mounty>(s, repository: repository);
         return await provider?.rawInsert(
-            'INSERT OR REPLACE INTO `_brick_KitchenSink_set_offline_first_model` (`KitchenSink_brick_id`, `Mounty_brick_id`) VALUES (?, ?)',
+            'INSERT OR IGNORE INTO `_brick_KitchenSink_set_offline_first_model` (`KitchenSink_brick_id`, `Mounty_brick_id`) VALUES (?, ?)',
             [instance.primaryKey, id]);
       }));
     }
 
     if (instance.primaryKey != null) {
-      await Future.wait<int>(
-          instance.futureListOfflineFirstModel?.map((s) async {
+      await Future.wait<int>(instance.futureListOfflineFirstModel?.map((s) async {
         final id = (await s)?.primaryKey ??
             await provider?.upsert<Mounty>((await s), repository: repository);
         return await provider?.rawInsert(
-            'INSERT OR REPLACE INTO `_brick_KitchenSink_future_list_offline_first_model` (`KitchenSink_brick_id`, `Mounty_brick_id`) VALUES (?, ?)',
+            'INSERT OR IGNORE INTO `_brick_KitchenSink_future_list_offline_first_model` (`KitchenSink_brick_id`, `Mounty_brick_id`) VALUES (?, ?)',
             [instance.primaryKey, id]);
       }));
     }
 
     if (instance.primaryKey != null) {
-      await Future.wait<int>(
-          instance.futureSetOfflineFirstModel?.map((s) async {
+      await Future.wait<int>(instance.futureSetOfflineFirstModel?.map((s) async {
         final id = (await s)?.primaryKey ??
             await provider?.upsert<Mounty>((await s), repository: repository);
         return await provider?.rawInsert(
-            'INSERT OR REPLACE INTO `_brick_KitchenSink_future_set_offline_first_model` (`KitchenSink_brick_id`, `Mounty_brick_id`) VALUES (?, ?)',
+            'INSERT OR IGNORE INTO `_brick_KitchenSink_future_set_offline_first_model` (`KitchenSink_brick_id`, `Mounty_brick_id`) VALUES (?, ?)',
             [instance.primaryKey, id]);
       }));
     }
   }
 
-  Future<KitchenSink> fromRest(Map<String, dynamic> input,
-          {provider, repository}) async =>
-      await _$KitchenSinkFromRest(input,
-          provider: provider, repository: repository);
-  Future<Map<String, dynamic>> toRest(KitchenSink input,
-          {provider, repository}) async =>
-      await _$KitchenSinkToRest(input,
-          provider: provider, repository: repository);
-  Future<KitchenSink> fromSqlite(Map<String, dynamic> input,
-          {provider, repository}) async =>
-      await _$KitchenSinkFromSqlite(input,
-          provider: provider, repository: repository);
-  Future<Map<String, dynamic>> toSqlite(KitchenSink input,
-          {provider, repository}) async =>
-      await _$KitchenSinkToSqlite(input,
-          provider: provider, repository: repository);
+  Future<KitchenSink> fromRest(Map<String, dynamic> input, {provider, repository}) async =>
+      await _$KitchenSinkFromRest(input, provider: provider, repository: repository);
+  Future<Map<String, dynamic>> toRest(KitchenSink input, {provider, repository}) async =>
+      await _$KitchenSinkToRest(input, provider: provider, repository: repository);
+  Future<KitchenSink> fromSqlite(Map<String, dynamic> input, {provider, repository}) async =>
+      await _$KitchenSinkFromSqlite(input, provider: provider, repository: repository);
+  Future<Map<String, dynamic>> toSqlite(KitchenSink input, {provider, repository}) async =>
+      await _$KitchenSinkToSqlite(input, provider: provider, repository: repository);
 }