+++ packages/brick_offline_first_with_rest_build/test/offline_first_schema_generator/test_with_associations.dart
@@ -15,19 +15,20 @@ import 'package:brick_sqlite_abstract/db.dart';
 import 'package:brick_sqlite_abstract/db.dart' show Migratable;
 
 /// All intelligently-generated migrations from all `@Migratable` classes on disk
-final Set<Migration> migrations = Set.from([]);
+final Set<Migration> migrations = <Migration>{};
 
 /// A consumable database structure including the latest generated migration.
 final schema = Schema(0,
     generatorVersion: 1,
-    tables: Set<SchemaTable>.from([
+    tables: <SchemaTable>{
       SchemaTable('SqliteAssoc',
-          columns: Set.from([
+          columns: <SchemaColumn>{
             SchemaColumn('_brick_id', int,
                 autoincrement: true, nullable: false, isPrimaryKey: true)
-          ])),
+          },
+          indices: <SchemaIndex>{}),
       SchemaTable('_brick_OneToOneAssocation_assocs',
-          columns: Set.from([
+          columns: <SchemaColumn>{
             SchemaColumn('_brick_id', int,
                 autoincrement: true, nullable: false, isPrimaryKey: true),
             SchemaColumn('l_OneToOneAssocation_brick_id', int,
@@ -40,9 +41,9 @@ final schema = Schema(0,
                 foreignTableName: 'SqliteAssoc',
                 onDeleteCascade: true,
                 onDeleteSetDefault: false)
-          ])),
+          }),
       SchemaTable('OneToOneAssocation',
-          columns: Set.from([
+          columns: <SchemaColumn>{
             SchemaColumn('_brick_id', int,
                 autoincrement: true, nullable: false, isPrimaryKey: true),
             SchemaColumn('assoc_SqliteAssoc_brick_id', int,
@@ -50,8 +51,9 @@ final schema = Schema(0,
                 foreignTableName: 'SqliteAssoc',
                 onDeleteCascade: false,
                 onDeleteSetDefault: false)
-          ]))
-    ]));
+          },
+          indices: <SchemaIndex>{})
+    });
 ''';
 
 @ConnectOfflineFirstWithRest()