+++ packages/brick_sqlite_abstract/lib/src/db/migration_commands/drop_index.dart
@@ -1,0 +1,14 @@
+import 'migration_command.dart';
+
+/// Drop index from DB if it exists
+class DropIndex extends MigrationCommand {
+  final String name;
+
+  const DropIndex(this.name);
+
+  @override
+  String get statement => 'DROP INDEX IF EXISTS $name';
+
+  @override
+  String get forGenerator => "DropIndex('$name')";
+}