+++ packages/brick_sqlite_abstract/test/db/__mocks__.dart
@@ -81,3 +81,23 @@ class MigrationInsertForeignKey extends Migration {
           down: const [],
         );
 }
+
+class MigrationCreateIndex extends Migration {
+  const MigrationCreateIndex()
+      : super(
+          version: 7,
+          up: const [
+            CreateIndex(columns: ['_brick_id'], onTable: 'demo', unique: true)
+          ],
+          down: const [DropIndex('index_demo_on__brick_id')],
+        );
+}
+
+class MigrationDropIndex extends Migration {
+  const MigrationDropIndex()
+      : super(
+          version: 8,
+          up: const [DropIndex('index_demo_on__brick_id')],
+          down: const [],
+        );
+}